nlp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
textstat==0.7.3
reportlab==4.1.0
slowapi==0.1.9
